moving grad app to components folder and update imports
